+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/internal/exporters/DiskExporter.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.contrib.disk.buffering.internal.exporters;
 
-import io.opentelemetry.contrib.disk.buffering.StoredBatchExporter;
 import io.opentelemetry.contrib.disk.buffering.internal.serialization.serializers.SignalSerializer;
 import io.opentelemetry.contrib.disk.buffering.internal.storage.Storage;
 import io.opentelemetry.contrib.disk.buffering.internal.storage.responses.ReadableResult;
