+++ disk-buffering/src/test/java/io/opentelemetry/contrib/disk/buffering/IntegrationTest.java
@@ -17,6 +17,10 @@ import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.contrib.disk.buffering.internal.StorageConfiguration;
+import io.opentelemetry.contrib.disk.buffering.internal.exporters.LogRecordDiskExporter;
+import io.opentelemetry.contrib.disk.buffering.internal.exporters.MetricDiskExporter;
+import io.opentelemetry.contrib.disk.buffering.internal.exporters.SpanDiskExporter;
+import io.opentelemetry.contrib.disk.buffering.internal.exporters.StoredBatchExporter;
 import io.opentelemetry.sdk.common.Clock;
 import io.opentelemetry.sdk.logs.SdkLoggerProvider;
 import io.opentelemetry.sdk.logs.export.LogRecordExporter;
